bug fix of assigning CRTM soil and vegetation types and add satellite ID 'n21' for FOV averaging  (#3835)
* Fix a bug that setts CRTM soil and vegetation types to be zeos with a number right under one.

* Add the ID "n21" for the satellite NOAA-N21 in calc_fov_crosstrk.f90.

* Fixing the case statement, to combine the case

---------

+++ src/ufo/operators/crtm/ufo_crtm_utils_mod.F90
@@ -1380,13 +1380,13 @@ real(kind_real), allocatable :: ObsTb(:,:)
     !Vegetation_Type
     call ufo_geovals_get_var(geovals, var_sfc_vegtyp, geoval)
     do k1 = 1, n_Profiles
-      sfc(k1)%Vegetation_Type = int(geoval%vals(1, k1))
+      sfc(k1)%Vegetation_Type = nint(geoval%vals(1, k1))
     end do
 
     !Soil_Type
     call ufo_geovals_get_var(geovals, var_sfc_soiltyp, geoval)
     do k1 = 1, n_Profiles
-      sfc(k1)%Soil_Type = int(geoval%vals(1, k1))
+      sfc(k1)%Soil_Type = nint(geoval%vals(1, k1))
     end do
   end if
